Correct handling of client ids
See the accompanying PR on verdi-runtime for more. Basically, in order
to do request deduplication we need clients to have the same id across
servers, and right now they almost definitely won't. The clients now
send their client id with every request, and vard is patched to work
with the new shim.

We should migrate these changes to the DiskOp shim as well. When
that's done, we can also change the other versions of vard that use
that shim.